+++ jgclark.QuickCapture/src/inbox.js
@@ -2,7 +2,7 @@
 // ----------------------------------------------------------------------------
 // Inbox command for QuickCapture plugin
 // by Jonathan Clark
-// last update 10.2.2024 for v0.16.0 by @jgclark
+// last update 2025-07-28 for v0.17.0 by @jgclark
 // ----------------------------------------------------------------------------
 
 import pluginJson from '../plugin.json'
@@ -13,7 +13,7 @@ import {
 } from './quickCaptureHelpers'
 import { logDebug, logError, logInfo, logWarn } from '@helpers/dev'
 import { displayTitle } from '@helpers/general'
-import { smartAppendPara, smartPrependPara } from '@helpers/paragraph'
+import { smartCreateSectionsAndPara } from '@helpers/paragraph'
 import {
   // chooseFolder, chooseHeading,
   showMessage
@@ -63,11 +63,11 @@ export async function addJotToInbox(
 
 /**
  * This adds a task to a special 'inbox' note. Possible configuration:
- * - add to the current Daily or Weekly note, or to a fixed note (through setting 'inboxLocation')
+ * - add to the current Daily or Weekly note, or to a fixed note (through setting 'inboxLocation') or to arg2 (if given)
  * - append or prepend to the inbox note (default: append)
  * Note: Internal function used by exported functions above.
  * @author @jgclark
- * @param {string?} itemType: 'task' (default) or 'jot'
+ * @param {string?} itemType: 'task' (default) or 'jot' (i.e. text)
  * @param {string?} taskArg (if not given, will ask user)
  * @param {string?} inboxTitleArg (if not given, will ask user)
  */
@@ -81,14 +81,22 @@ async function addItemToInbox(
     const paraType = itemType === 'task' ? 'open' : 'text' 
     const config = await getQuickCaptureSettings()
     logDebug(pluginJson, `addItemToInbox(): starting for ${itemType} (= paraType ${paraType}) with ${config.inboxLocation}`)
+    const textToAppend = (config.textToAppendToTasks && itemType === 'task')
+      ? ` ${config.textToAppendToTasks}`
+      : (config.textToAppendToJots && itemType === 'jot')
+        ? ` ${config.textToAppendToJots}`
+        : ''
+    const inboxHeading = (config.inboxHeading && config.inboxHeading !== '') ? config.inboxHeading : ''
 
     // TEST: Extra possible arg
     // let inboxNote: ?TNote
     let inboxTitleToUse = ''
-    if (!inboxTitleArg || inboxTitleArg === '') {
+    if (inboxTitleArg && inboxTitleArg !== '') {
+      inboxTitleToUse = inboxTitleArg
+      logDebug('addItemToInbox', `Title arg given: inboxTitleToUse=${inboxTitleToUse}`)
+    } else {
       switch (config.inboxLocation) {
         case "Daily": {
-          // inboxNote = DataStore.calendarNoteByDate(new Date(), "day")
           inboxTitleToUse = 'today'
           break
         }
@@ -99,7 +107,7 @@ async function addItemToInbox(
         }
 
         default: {
-          if (config.inboxTitle === '') {
+          if (!config.inboxTitle || config.inboxTitle === '') {
             throw new Error("Quick Capture to Inbox: please set the title of your chosen fixed Inbox note in Quick Capture preferences.")
           } else {
             inboxTitleToUse = config.inboxTitle
@@ -108,13 +116,9 @@ async function addItemToInbox(
         }
       }
       logDebug('addItemToInbox', `No title arg given: inboxTitleToUse=${inboxTitleToUse}`)
-    } else {
-      inboxTitleToUse = inboxTitleArg
-      logDebug('addItemToInbox', `Title arg given: inboxTitleToUse=${inboxTitleToUse}`)
     }
 
-    const inboxNote = await getNoteFromParamOrUser(`Inbox ${itemType}`, inboxTitleToUse, false)
-
+    const inboxNote = await getNoteFromParamOrUser(`Inbox ${itemType}`, inboxTitleToUse)
     if (!inboxNote) {
       throw new Error("Quick Add to Inbox: Couldn't get or make valid Inbox note.")
     }
@@ -122,20 +126,18 @@ async function addItemToInbox(
     // Get item title either from passed argument or ask user
     let itemText = (itemArg != null && itemArg !== '')
       ? itemArg
-      : await CommandBar.showInput(`Type the ${itemType} to add to ${displayTitle(inboxNote)}`, `Add ${itemType} '%@' ${config.textToAppendToTasks}`)
-    if (itemType === 'task') {
-      itemText += ` ${config.textToAppendToTasks}`
-    } else if (itemType === 'jot') {
-      itemText += ` ${config.textToAppendToJots}`
+      : await CommandBar.showInput(`Type the ${itemType} to add to ${displayTitle(inboxNote)}`, `Add ${itemType} '%@'${textToAppend}`)
+    if (itemType === 'jot') {
+      itemText += textToAppend
     }
 
     if (config.addInboxPosition === 'append') {
       // inboxNote.appendTodo(itemText)
-      smartAppendPara(inboxNote, itemText, paraType)
+      smartCreateSectionsAndPara(inboxNote, itemText, paraType, [inboxHeading], config.headingLevel, true)
       logDebug(pluginJson, `- appended to note '${displayTitle(inboxNote)}'`)
     } else {
       // inboxNote.prependTodo(itemText)
-      smartPrependPara(inboxNote, itemText, paraType)
+      smartCreateSectionsAndPara(inboxNote, itemText, paraType, [inboxHeading], config.headingLevel, false)
       logDebug(pluginJson, `- prepended to note '${displayTitle(inboxNote)}'`)
     }
   }